Add gif generator template
