--- db/Updater/descriptor_generator_test/temp/contour.inv
@@ -1,122 +1,0 @@
-121
-7.91117
-1.31776e-15
-10.5755
-10.6483
-0.327797
-1.34927
-5.28531
-5.77726
-5.803
-0.552045
-2.88601
-2.95674
-0.142617
-1.39631
-1.87453
-1.92961
-0.161485
-2.16452
-2.92388
-2.94056
-1.89521
-2.5003
-2.70412
-3.03245
-3.04298
-0.461316
-1.5406
-2.92102
-3.14072
-3.14689
-1.5701
-2.35813
-2.67799
-3.06129
-3.21349
-3.21668
-0.526431
-1.46106
-2.33373
-2.98821
-3.09598
-3.09794
-0.623466
-1.22267
-1.92787
-2.21668
-2.72629
-2.80383
-2.80455
-0.450464
-1.29705
-1.78917
-2.58075
-2.92576
-2.97567
-2.97606
-0.295772
-0.772289
-1.39412
-2.10952
-2.66
-2.95806
-2.9746
-2.97489
-0.348986
-0.615976
-0.853772
-1.94024
-2.3554
-2.6329
-2.64579
-2.64594
-0.33256
-0.591242
-1.42745
-2.16717
-2.78496
-3.13088
-3.2045
-3.21294
-3.21299
-0.246015
-0.867108
-1.13849
-2.10261
-2.4954
-2.90896
-2.97755
-2.98171
-2.98177
-0.277576
-0.428553
-1.00529
-1.64085
-2.3155
-2.56846
-2.73357
-2.78992
-2.7922
-2.79225
-0.569398
-1.14386
-1.3925
-1.8595
-2.42459
-2.7599
-2.91904
-2.94475
-2.9469
-2.94692
-0.487787
-0.64399
-0.809161
-1.17684
-1.5539
-1.97927
-2.24766
-2.37362
-2.39383
-2.39616
-2.39618
