--- db/Updater/descriptor_generator_test/temp/one_std_contour.inv
@@ -1,242 +1,0 @@
-7.91117
-1.31776e-15
-10.5755
-10.6483
-0.327797
-1.34927
-5.28531
-5.77726
-5.803
-0.552045
-2.88601
-2.95674
-0.142617
-1.39631
-1.87453
-1.92961
-0.161485
-2.16452
-2.92388
-2.94056
-1.89521
-2.5003
-2.70412
-3.03245
-3.04298
-0.461316
-1.5406
-2.92102
-3.14072
-3.14689
-1.5701
-2.35813
-2.67799
-3.06129
-3.21349
-3.21668
-0.526431
-1.46106
-2.33373
-2.98821
-3.09598
-3.09794
-0.623466
-1.22267
-1.92787
-2.21668
-2.72629
-2.80383
-2.80455
-0.450464
-1.29705
-1.78917
-2.58075
-2.92576
-2.97567
-2.97606
-0.295772
-0.772289
-1.39412
-2.10952
-2.66
-2.95806
-2.9746
-2.97489
-0.348986
-0.615976
-0.853772
-1.94024
-2.3554
-2.6329
-2.64579
-2.64594
-0.33256
-0.591242
-1.42745
-2.16717
-2.78496
-3.13088
-3.2045
-3.21294
-3.21299
-0.246015
-0.867108
-1.13849
-2.10261
-2.4954
-2.90896
-2.97755
-2.98171
-2.98177
-0.277576
-0.428553
-1.00529
-1.64085
-2.3155
-2.56846
-2.73357
-2.78992
-2.7922
-2.79225
-0.569398
-1.14386
-1.3925
-1.8595
-2.42459
-2.7599
-2.91904
-2.94475
-2.9469
-2.94692
-0.487787
-0.64399
-0.809161
-1.17684
-1.5539
-1.97927
-2.24766
-2.37362
-2.39383
-2.39616
-2.39618
-6.37598
-1.19532e-15
-8.52337
-8.57696
-0.174058
-1.04013
-4.28174
-4.60215
-4.62185
-0.166749
-2.11561
-2.18314
-0.136279
-1.17031
-1.50588
-1.55593
-0.346175
-1.5554
-2.21155
-2.22959
-1.65958
-2.21587
-2.39716
-2.62829
-2.63853
-0.614756
-1.13267
-2.23095
-2.44252
-2.44904
-1.31682
-1.93988
-2.30947
-2.49068
-2.62919
-2.63238
-0.486137
-1.06449
-1.85397
-2.3702
-2.46952
-2.47183
-0.34403
-1.02471
-1.61099
-1.84357
-2.23025
-2.29776
-2.29865
-0.368856
-1.0912
-1.48697
-2.03612
-2.30429
-2.35659
-2.35716
-0.282794
-0.628446
-1.2826
-1.89453
-2.32342
-2.51861
-2.53701
-2.53738
-0.194851
-0.432452
-0.87475
-1.69572
-1.9645
-2.18797
-2.20409
-2.20426
-0.0315104
-0.522797
-1.27255
-1.72597
-2.34532
-2.52275
-2.60212
-2.61197
-2.61205
-0.246474
-0.887434
-1.17486
-1.83562
-2.20225
-2.43703
-2.50848
-2.51287
-2.51295
-0.525877
-0.607766
-0.907995
-1.24862
-1.78913
-2.01867
-2.17621
-2.23364
-2.23669
-2.23675
-0.264208
-0.890554
-1.07159
-1.46183
-1.9997
-2.19566
-2.33198
-2.3579
-2.36077
-2.3608
-0.397236
-0.557088
-0.747032
-1.04926
-1.34275
-1.7937
-1.99482
-2.09403
-2.11409
-2.11658
-2.1166
