--- db/Updater/descriptor_generator_test/temp/one_third_two_thirds_contour.inv
@@ -1,363 +1,0 @@
-7.91117
-1.31776e-15
-10.5755
-10.6483
-0.327797
-1.34927
-5.28531
-5.77726
-5.803
-0.552045
-2.88601
-2.95674
-0.142617
-1.39631
-1.87453
-1.92961
-0.161485
-2.16452
-2.92388
-2.94056
-1.89521
-2.5003
-2.70412
-3.03245
-3.04298
-0.461316
-1.5406
-2.92102
-3.14072
-3.14689
-1.5701
-2.35813
-2.67799
-3.06129
-3.21349
-3.21668
-0.526431
-1.46106
-2.33373
-2.98821
-3.09598
-3.09794
-0.623466
-1.22267
-1.92787
-2.21668
-2.72629
-2.80383
-2.80455
-0.450464
-1.29705
-1.78917
-2.58075
-2.92576
-2.97567
-2.97606
-0.295772
-0.772289
-1.39412
-2.10952
-2.66
-2.95806
-2.9746
-2.97489
-0.348986
-0.615976
-0.853772
-1.94024
-2.3554
-2.6329
-2.64579
-2.64594
-0.33256
-0.591242
-1.42745
-2.16717
-2.78496
-3.13088
-3.2045
-3.21294
-3.21299
-0.246015
-0.867108
-1.13849
-2.10261
-2.4954
-2.90896
-2.97755
-2.98171
-2.98177
-0.277576
-0.428553
-1.00529
-1.64085
-2.3155
-2.56846
-2.73357
-2.78992
-2.7922
-2.79225
-0.569398
-1.14386
-1.3925
-1.8595
-2.42459
-2.7599
-2.91904
-2.94475
-2.9469
-2.94692
-0.487787
-0.64399
-0.809161
-1.17684
-1.5539
-1.97927
-2.24766
-2.37362
-2.39383
-2.39616
-2.39618
-2.02793
-4.29774e-16
-2.71134
-2.73121
-0.251926
-0.337163
-1.38645
-1.57738
-1.58201
-0.522366
-0.777274
-0.788588
-0.0475016
-0.90062
-0.963902
-0.967299
-0.438878
-0.837
-0.922057
-0.924383
-0.303659
-0.82061
-0.945988
-0.975371
-0.976259
-0.28236
-0.670089
-0.841307
-0.870986
-0.871357
-0.108893
-0.556103
-0.687814
-0.772673
-0.783591
-0.783785
-0.284925
-0.525113
-0.691309
-0.794744
-0.79994
-0.800153
-0.0382055
-0.308781
-0.448596
-0.5967
-0.635477
-0.63881
-0.638983
-0.322337
-0.423287
-0.558712
-0.723375
-0.745885
-0.749667
-0.749753
-0.0273652
-0.174952
-0.341425
-0.477715
-0.55349
-0.57432
-0.577427
-0.577548
-0.296324
-0.366662
-0.468284
-0.613412
-0.659902
-0.677233
-0.67849
-0.678579
-0.0200719
-0.187576
-0.289573
-0.406999
-0.481241
-0.539384
-0.550491
-0.552208
-0.55231
-0.222914
-0.301842
-0.396607
-0.491306
-0.547849
-0.586449
-0.591339
-0.592652
-0.592763
-0.0599332
-0.125408
-0.227091
-0.349184
-0.399276
-0.477074
-0.494451
-0.501241
-0.502612
-0.50272
-0.104804
-0.186111
-0.291088
-0.378459
-0.436869
-0.49737
-0.510202
-0.515668
-0.517408
-0.517515
-0.0482044
-0.106842
-0.179685
-0.253766
-0.330308
-0.391859
-0.429086
-0.442286
-0.448384
-0.449936
-0.450044
-0.284548
-8.48044e-17
-0.380605
-0.389782
-0.0519743
-0.0805395
-0.220473
-0.295423
-0.296289
-0.119916
-0.183466
-0.18408
-0.0617021
-0.237273
-0.242501
-0.242925
-0.135914
-0.196425
-0.201356
-0.201899
-0.00925124
-0.171868
-0.19132
-0.196654
-0.19693
-0.104044
-0.148584
-0.161164
-0.164404
-0.16464
-0.0126859
-0.0882688
-0.138847
-0.156462
-0.158126
-0.158521
-0.056085
-0.109396
-0.126166
-0.13562
-0.136582
-0.136921
-0.02332
-0.0600576
-0.108427
-0.127796
-0.134548
-0.135768
-0.136081
-0.0219772
-0.0935481
-0.122904
-0.135676
-0.140774
-0.141617
-0.141936
-0.03025
-0.0574899
-0.0824715
-0.0947426
-0.110309
-0.113818
-0.114546
-0.114971
-0.0289731
-0.099699
-0.129301
-0.143363
-0.152925
-0.155859
-0.156593
-0.156857
-0.0323856
-0.0448405
-0.0549418
-0.0724627
-0.0858018
-0.0945859
-0.0976201
-0.0985542
-0.098876
-0.0485735
-0.101947
-0.124923
-0.141409
-0.151779
-0.158688
-0.160298
-0.160818
-0.161071
-0.027977
-0.0382181
-0.0580199
-0.0979015
-0.108604
-0.11852
-0.127985
-0.129936
-0.130385
-0.130682
-0.0427175
-0.081067
-0.10807
-0.129458
-0.147506
-0.161462
-0.164046
-0.165061
-0.165444
-0.165629
-0.00256042
-0.0362082
-0.0621339
-0.121748
-0.137815
-0.151764
-0.163147
-0.167475
-0.168609
-0.169015
-0.169231
