+++ db/Updater/updater/Main.py
@@ -5,11 +5,12 @@ from classes.Update import Update
 from datetime import date
 from classes.Time_stamp import Time_stamp
 from psycopg2 import sql
-from descriptors_generator.values_generator import get_emd_descriptors
+from generators.Values_generator import get_emd_descriptors, remove_map, download_file
 from classes.Type_descriptor import Type_descriptor
 from classes.Descriptor import Descriptor
-from descriptors_generator import values_generator as valg
+from generators import Values_generator as valg
 from time import time 
+from generators import Gif_generator as gifG
  
 '''
 Created on 31 mar. 2019
@@ -17,7 +18,8 @@ Created on 31 mar. 2019
 @author: luis98
 '''
 
-valg.dir = "../descriptors_generator/"
+valg.dir = "../generators/"
+gifG.dir = "../generators/"
 
 def create_type_descriptors(cursor_sql):
     
@@ -51,16 +53,22 @@ def update_descriptor(emd_entry_p,cursor_sql):
         des_temp.update_db(cursor_sql)
 
 
-def first_time_emd(emds,conec_ftp,conec_sql,emd_url, emd_url_directory):
+def first_time_emd(conec_ftp,conec_sql,emd_url_server, emd_url_path):
+    emds = conec_ftp.get_all_emds_id()
     cursor_sql = conec_sql.get_cursor()
     
     create_type_descriptors(cursor_sql)
     print("Total inserts",len(emds))
     for i in emds:
         temp_emd_entry = Emd_entry()
+        temp_emd_entry.emd_url = "http://{0}{1}".format(emd_url_server,emd_url_path)
         temp_emd_entry.create_by_ftp(i,conec_ftp.ftp)
         temp_emd_entry.insert_db(cursor_sql)
+        
+        download_file(i)
+        gifG.generateGif(i)
         insert_descriptor(temp_emd_entry,cursor_sql)
+        remove_map(i)
         
         temp_time_stamp = Time_stamp (i,date.today(),temp_emd_entry.map_url,temp_emd_entry.xml_url,temp_emd_entry.image_url )
         temp_time_stamp.insert_db(cursor_sql)
@@ -72,13 +80,14 @@ def first_time_emd(emds,conec_ftp,conec_sql,emd_url, emd_url_directory):
     conec_sql.commit()
     cursor_sql.close()
     
-def update_emd(conec_ftp,conec_sql):
+def update_emd(conec_ftp,conec_sql,emd_url_server, emd_url_path):
     cursor_sql = conec_sql.get_cursor()
     
     emds = conec_ftp.get_emds_higher_than_date(conec_sql.last_update())
     print("Total changes",len(emds))
     for i in emds:
         temp_emd_entry = Emd_entry()
+        temp_emd_entry.emd_url = "http://{0}{1}".format(emd_url_server,emd_url_path)
         temp_emd_entry.create_by_ftp(i,conec_ftp.ftp)
         temp_time_stamp = Time_stamp (i,date.today(),temp_emd_entry.map_url,temp_emd_entry.xml_url,temp_emd_entry.image_url )
         
@@ -89,11 +98,17 @@ def update_emd(conec_ftp,conec_sql):
             temp_emd_entry.map_id = map_result[0]
             temp_emd_entry.update_db(cursor_sql)
             temp_time_stamp.update_db(cursor_sql)
+            download_file(i)
+            gifG.generateGif(i)
             update_descriptor(temp_emd_entry,cursor_sql)
+            remove_map(i)
         else:
             temp_emd_entry.insert_db(cursor_sql)
             temp_time_stamp.insert_db(cursor_sql)
+            download_file(i)
+            gifG.generateGif(i)
             insert_descriptor(temp_emd_entry,cursor_sql)
+            remove_map(i)
         
         conec_sql.commit()
         print(i)
@@ -106,19 +121,17 @@ def update_emd(conec_ftp,conec_sql):
     cursor_sql.close()
     
 
-def main(emd_url, emd_url_directory, num_of_nucles):
-    conec_ftp = FTP_connection()
-    conec_ftp.init_connection()
+def main(emd_url_server, emd_url_path):
+    valg.emd_url = "http://{0}{1}".format(emd_url_server,emd_url_path)
+    conec_ftp = FTP_connection(ftp_server=emd_url_server)
+    conec_ftp.init_connection(initial_directory="{0}{1}".format(emd_url_path,"/structures/"))
     conec_sql = SQL_connection()
     conec_sql.init_connection()
     
     if conec_sql.is_first_time():
-        emds = conec_ftp.get_all_emds_id()
-        total_emd = len(emds)
-        
-        first_time_emd(emds, conec_ftp,conec_sql,emd_url, emd_url_directory)
+        first_time_emd(conec_ftp,conec_sql,emd_url_server, emd_url_path)
     else:
-        update_emd(emds, conec_ftp,conec_sql,emd_url, emd_url_directory)
+        update_emd(conec_ftp,conec_sql,emd_url_server, emd_url_path)
         
     conec_sql.close_connection()    
     conec_ftp.close_connection()
@@ -126,7 +139,7 @@ def main(emd_url, emd_url_directory, num_of_nucles):
   
 if __name__== "__main__":
     ini = time() 
-    main()
+    main("ftp.pdbj.org","/pub/emdb")
     final = time() 
     ejec = final - ini
     print ('Execution time:', ejec)
